# tests/contract/api/openapi_snapshot.json
{
 "components": {
  "schemas": {
...
   "CreateRunRequest": {
    "additionalProperties": false,
    "description": "POST /runs \u7684\u8bf7\u6c42\u4f53\uff08planner \u610f\u56fe\uff09\u3002",
    "properties": {
     "template": {
      "default": "single-fixture",
      "maxLength": 64,
      "title": "Template",
      "type": "string"
     },
     "workspace_id": {
      "format": "uuid",
      "title": "Workspace Id",
      "type": "string"
     }
    },
    "required": [
     "workspace_id"
    ],
    "title": "CreateRunRequest",
    "type": "object"
   },
...
   "EvalRunDetailView": {
    "additionalProperties": false,
    "properties": {
     "eval_run_id": {
      "title": "Eval Run Id",
      "type": "string"
     },
     "mode": {
      "title": "Mode",
      "type": "string"
     },
     "outcomes": {
      "items": {
       "$ref": "#/components/schemas/EvalOutcomeView"
      },
      "title": "Outcomes",
      "type": "array"
     },
     "registered_units": {
      "items": {
       "$ref": "#/components/schemas/RegisteredUnit"
      },
      "title": "Registered Units",
      "type": "array"
     },
     "report": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Report"
     },
     "run_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run Id"
     },
     "sealed_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sealed At"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "status_breakdown": {
      "$ref": "#/components/schemas/EvalRunStatusBreakdown"
     }
    },
    "required": [
     "eval_run_id",
     "run_id",
     "mode",
     "status",
     "sealed_at",
     "registered_units",
     "outcomes",
     "status_breakdown"
    ],
    "title": "EvalRunDetailView",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "RunDetail": {
    "additionalProperties": false,
    "description": "run \u8be6\u60c5\uff08\u542b\u4efb\u52a1\u6295\u5f71\uff09\u3002",
    "properties": {
     "mode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mode"
     },
     "organization_id": {
      "format": "uuid",
      "title": "Organization Id",
      "type": "string"
     },
     "run_id": {
      "format": "uuid",
      "title": "Run Id",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "tasks": {
      "additionalProperties": {
       "additionalProperties": true,
       "type": "object"
      },
      "default": {},
      "title": "Tasks",
      "type": "object"
     },
     "template": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template"
     }
    },
    "required": [
     "run_id",
     "status",
     "organization_id"
    ],
    "title": "RunDetail",
    "type": "object"
   },
...
   "RunRecord": {
    "additionalProperties": false,
    "description": "\u4e00\u4e2a run \u7684\u5217\u8868\u9879\u6295\u5f71\uff08\u6765\u81ea PG reduce\uff09\u3002",
    "properties": {
     "organization_id": {
      "format": "uuid",
      "title": "Organization Id",
      "type": "string"
     },
     "run_id": {
      "format": "uuid",
      "title": "Run Id",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     },
     "template": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template"
     }
    },
    "required": [
     "run_id",
     "status",
     "organization_id"
    ],
    "title": "RunRecord",
    "type": "object"
   },
   "SealEvalRunRequest": {
    "additionalProperties": false,
    "description": "seal \u8bf7\u6c42\u4f53\u4e0e EvalFoundationService.seal \u7b7e\u540d\u9010\u53c2\u6570\u5bf9\u9f50\uff0c\u4e0d\u5f31\u5316\uff1a\nmigration_revision \u4e0e test_report \u90fd\u662f\u5fc5\u9700\u8f93\u5165\uff08\u5bc6\u5c01\u4ef6\u5f15\u7528\u8fc1\u79fb\u57fa\u7ebf\u4e0e\nS0 eval \u5951\u7ea6\u6d4b\u8bd5\u8bc1\u636e\uff0c\u7f3a\u4efb\u4e00\u90fd\u65e0\u6cd5\u6784\u6210\u53ef\u590d\u6838\u7684\u5bc6\u5c01\uff09\u3002",
    "properties": {
     "migration_revision": {
      "title": "Migration Revision",
      "type": "string"
     },
     "test_report": {
      "additionalProperties": true,
      "title": "Test Report",
      "type": "object"
     }
    },
    "required": [
     "migration_revision",
     "test_report"
    ],
    "title": "SealEvalRunRequest",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "zhiwei",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/evals/{eval_run_id}": {
   "get": {
    "operationId": "get_eval_run_api_v1_evals__eval_run_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "eval_run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Eval Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EvalRunDetailView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Eval Run",
    "tags": [
     "evals"
    ]
   }
  },
  "/api/v1/evals/{eval_run_id}/report": {
   "get": {
    "operationId": "get_eval_run_report_api_v1_evals__eval_run_id__report_get",
    "parameters": [
     {
      "in": "path",
      "name": "eval_run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Eval Run Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "model",
      "required": true,
      "schema": {
       "title": "Model",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "version",
      "required": true,
      "schema": {
       "title": "Version",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "date",
      "required": true,
      "schema": {
       "title": "Date",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "corpus",
      "required": true,
      "schema": {
       "title": "Corpus",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "environment",
      "required": true,
      "schema": {
       "title": "Environment",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Eval Run Report Api V1 Evals  Eval Run Id  Report Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Eval Run Report",
    "tags": [
     "evals"
    ]
   }
  },
  "/api/v1/evals/{eval_run_id}/resume": {
   "post": {
    "operationId": "resume_eval_run_api_v1_evals__eval_run_id__resume_post",
    "parameters": [
     {
      "in": "path",
      "name": "eval_run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Eval Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EvalRunDetailView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Resume Eval Run",
    "tags": [
     "evals"
    ]
   }
  },
  "/api/v1/evals/{eval_run_id}/seal": {
   "post": {
    "operationId": "seal_eval_run_api_v1_evals__eval_run_id__seal_post",
    "parameters": [
     {
      "in": "path",
      "name": "eval_run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Eval Run Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SealEvalRunRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EvalRunDetailView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Seal Eval Run",
    "tags": [
     "evals"
    ]
   }
  },
...
  "/api/v1/runs": {
   "get": {
    "operationId": "list_runs_api_v1_runs_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/RunRecord"
         },
         "title": "Response List Runs Api V1 Runs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Runs",
    "tags": [
     "runs"
    ]
   },
   "post": {
    "operationId": "create_run_api_v1_runs_post",
    "parameters": [
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "minLength": 1,
         "pattern": "\\S+",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateRunRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Idempotent replay (Idempotency-Key \u91cd\u8bd5\u547d\u4e2d\u5df2\u521b\u5efa Run)"
     },
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Run",
    "tags": [
     "runs"
    ]
   }
  },
  "/api/v1/runs/{run_id}": {
   "get": {
    "operationId": "get_run_api_v1_runs__run_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run",
    "tags": [
     "runs"
    ]
   }
  },
...
 }
}